+++ src/main/java/org/thema/pixscape/view/cuda/ComputeViewCUDA.java
@@ -88,7 +88,7 @@ public class ComputeViewCUDA extends SimpleComputeView {
     }
 
     @Override
-    public double calcRay(GridCoordinates2D c0, double startZ, GridCoordinates2D c1, double destZ, Bounds bounds) {
+    public double calcRay(GridCoordinates2D c0, double startZ, GridCoordinates2D c1, double destZ, Bounds bounds, boolean area) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
     
@@ -121,7 +121,7 @@ public class ComputeViewCUDA extends SimpleComputeView {
     }
 
     @Override
-    public ViewShedResult calcViewShedDeg(DirectPosition2D cg, double startZ, double destZ, boolean inverse, Bounds bounds) {
+    public ViewShedResult calcViewShedDeg(DirectPosition2D cg, double startZ, double destZ, boolean inverse, Bounds bounds, boolean area) {
         throw new UnsupportedOperationException("Degree viewshed is not supported on CUDA"); 
     }